some refactorings
